# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: The Jobs MV Phase 1 API
servers:
  - url: /api
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      security:
        - ClerkCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
...
  /candidate/cv/upload-url:
    post:
      operationId: requestCandidateCvUploadUrl
      security:
        - ClerkCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateCvUploadRequest"
      responses:
        "200":
          description: Private URL for directly uploading the candidate's CV
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateCvUploadResponse"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /candidate/cv:
    get:
      operationId: downloadCandidateCv
      security:
        - ClerkCookieAuth: []
      responses:
        "200":
          description: Candidate CV
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/msword:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/Error"
  /candidate/cv/complete:
    post:
      operationId: completeCandidateCvUpload
      security:
        - ClerkCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateCvUploadCompleteRequest"
      responses:
        "200":
          description: CV upload was verified and activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateCvUploadCompleteResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    ClerkCookieAuth:
      type: apiKey
      in: cookie
      name: __session
    ClerkBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    UploadUrlRequest:
      type: object
      required:
        - name
        - size
        - contentType
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
    UploadUrlResponse:
      type: object
      required:
        - uploadURL
        - objectPath
        - metadata
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
...
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    CandidateCvUploadRequest:
      type: object
      required:
        - name
        - size
        - contentType
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        size:
          type: integer
          minimum: 1
          maximum: 10485760
        contentType:
          type: string
          enum:
            - application/pdf
            - application/msword
            - application/vnd.openxmlformats-officedocument.wordprocessingml.document
    CandidateCvUploadResponse:
      type: object
      required:
        - uploadURL
        - uploadToken
        - expiresAt
      properties:
        uploadURL:
          type: string
          format: uri
        uploadToken:
          type: string
        expiresAt:
          type: string
          format: date-time
    CandidateCvUploadCompleteRequest:
      type: object
      required:
        - uploadToken
      properties:
        uploadToken:
          type: string
          minLength: 1
          maxLength: 4096
    CandidateCvUploadCompleteResponse:
      type: object
      required:
        - cvAvailable
      properties:
        cvAvailable:
          type: boolean
...