# lib/api-zod/src/generated/types/candidateCvUploadResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Jobs MV Phase 1 API
 * OpenAPI spec version: 0.1.0
 */

export interface CandidateCvUploadResponse {
  uploadURL: string;
  uploadToken: string;
  expiresAt: Date;
}